+++ src/server/models/server-settings/server-settings.ts
@@ -17,7 +17,8 @@
 import { Class, Instance, isInstanceOf } from 'immutable-class';
 
 export type Iframe = "allow" | "deny";
-export type StrictTransportSecurity = "none" | "when-x-forwarded-proto" | "always";
+export type TrustProxy = "none" | "always";
+export type StrictTransportSecurity = "none" | "always";
 
 export interface ServerSettingsValue {
   port?: number;
@@ -25,6 +26,7 @@ export interface ServerSettingsValue {
   serverRoot?: string;
   pageMustLoadTimeout?: number;
   iframe?: Iframe;
+  trustProxy?: TrustProxy;
   strictTransportSecurity?: StrictTransportSecurity;
 }
 
@@ -34,11 +36,19 @@ export interface ServerSettingsJS {
   serverRoot?: string;
   pageMustLoadTimeout?: number;
   iframe?: Iframe;
+  trustProxy?: TrustProxy;
   strictTransportSecurity?: StrictTransportSecurity;
 }
 
 function parseIntFromPossibleString(x: any) {
   return typeof x === 'string' ? parseInt(x, 10) : x;
+}
+
+function ensureOneOfOrNull<T>(name: string, thing: T, things: T[]): void {
+  if (thing == null) return;
+  if (things.indexOf(thing) === -1) {
+    throw new Error(`'${thing}' is not a valid value for ${name}, must be one of: ${things.join(', ')}`);
+  }
 }
 
 var check: Class<ServerSettingsValue, ServerSettingsJS>;
@@ -48,7 +58,9 @@ export class ServerSettings implements Instance<ServerSettingsValue, ServerSetti
   static DEFAULT_PAGE_MUST_LOAD_TIMEOUT = 800;
   static IFRAME_VALUES: Iframe[] = ["allow", "deny"];
   static DEFAULT_IFRAME: Iframe = "allow";
-  static STRICT_TRANSPORT_SECURITY_VALUES: StrictTransportSecurity[] = ["none", "when-x-forwarded-proto", "always"];
+  static TRUST_PROXY_VALUES: TrustProxy[] = ["none", "always"];
+  static DEFAULT_TRUST_PROXY: TrustProxy = "none";
+  static STRICT_TRANSPORT_SECURITY_VALUES: StrictTransportSecurity[] = ["none", "always"];
   static DEFAULT_STRICT_TRANSPORT_SECURITY: StrictTransportSecurity = "none";
 
   static isServerSettings(candidate: any): candidate is ServerSettings {
@@ -62,6 +74,7 @@ export class ServerSettings implements Instance<ServerSettingsValue, ServerSetti
       serverRoot,
       pageMustLoadTimeout,
       iframe,
+      trustProxy,
       strictTransportSecurity
     } = parameters;
 
@@ -74,6 +87,7 @@ export class ServerSettings implements Instance<ServerSettingsValue, ServerSetti
       serverRoot,
       pageMustLoadTimeout,
       iframe,
+      trustProxy,
       strictTransportSecurity
     });
   }
@@ -83,6 +97,7 @@ export class ServerSettings implements Instance<ServerSettingsValue, ServerSetti
   public serverRoot: string;
   public pageMustLoadTimeout: number;
   public iframe: Iframe;
+  public trustProxy: TrustProxy;
   public strictTransportSecurity: StrictTransportSecurity;
   public druidRequestDecorator: string;
 
@@ -94,8 +109,15 @@ export class ServerSettings implements Instance<ServerSettingsValue, ServerSetti
     this.serverHost = parameters.serverHost;
     this.serverRoot = parameters.serverRoot;
     this.pageMustLoadTimeout = parameters.pageMustLoadTimeout;
+
     this.iframe = parameters.iframe;
+    ensureOneOfOrNull('iframe', this.iframe, ServerSettings.IFRAME_VALUES);
+
+    this.trustProxy = parameters.trustProxy;
+    ensureOneOfOrNull('trustProxy', this.trustProxy, ServerSettings.TRUST_PROXY_VALUES);
+
     this.strictTransportSecurity = parameters.strictTransportSecurity;
+    ensureOneOfOrNull('strictTransportSecurity', this.strictTransportSecurity, ServerSettings.STRICT_TRANSPORT_SECURITY_VALUES);
   }
 
   public valueOf(): ServerSettingsValue {
@@ -105,6 +127,7 @@ export class ServerSettings implements Instance<ServerSettingsValue, ServerSetti
       serverRoot: this.serverRoot,
       pageMustLoadTimeout: this.pageMustLoadTimeout,
       iframe: this.iframe,
+      trustProxy: this.trustProxy,
       strictTransportSecurity: this.strictTransportSecurity
     };
   }
@@ -117,6 +140,7 @@ export class ServerSettings implements Instance<ServerSettingsValue, ServerSetti
     if (this.serverRoot) js.serverRoot = this.serverRoot;
     if (this.pageMustLoadTimeout) js.pageMustLoadTimeout = this.pageMustLoadTimeout;
     if (this.iframe) js.iframe = this.iframe;
+    if (this.trustProxy) js.trustProxy = this.trustProxy;
     if (this.strictTransportSecurity) js.strictTransportSecurity = this.strictTransportSecurity;
     return js;
   }
@@ -136,6 +160,7 @@ export class ServerSettings implements Instance<ServerSettingsValue, ServerSetti
       this.serverRoot === other.serverRoot &&
       this.pageMustLoadTimeout === other.pageMustLoadTimeout &&
       this.iframe === other.iframe &&
+      this.trustProxy === other.trustProxy &&
       this.strictTransportSecurity === other.strictTransportSecurity;
   }
 
@@ -153,6 +178,10 @@ export class ServerSettings implements Instance<ServerSettingsValue, ServerSetti
 
   public getIframe(): Iframe {
     return this.iframe || ServerSettings.DEFAULT_IFRAME;
+  }
+
+  public getTrustProxy(): TrustProxy {
+    return this.trustProxy || ServerSettings.DEFAULT_TRUST_PROXY;
   }
 
   public getStrictTransportSecurity(): StrictTransportSecurity {